+++ src/Mock/Handle/InstanceHandle.php
@@ -16,6 +16,7 @@ use Eloquent\Phony\Assertion\AssertionRenderer;
 use Eloquent\Phony\Call\Arguments;
 use Eloquent\Phony\Invocation\Invoker;
 use Eloquent\Phony\Mock\Mock;
+use Eloquent\Phony\Stub\EmptyValueFactory;
 use Eloquent\Phony\Stub\StubFactory;
 use Eloquent\Phony\Stub\StubVerifierFactory;
 use ReflectionClass;
@@ -34,6 +35,7 @@ class InstanceHandle extends AbstractHandle
      * @param stdClass            $state               The state.
      * @param StubFactory         $stubFactory         The stub factory to use.
      * @param StubVerifierFactory $stubVerifierFactory The stub verifier factory to use.
+     * @param EmptyValueFactory   $emptyValueFactory   The empty value factory to use.
      * @param AssertionRenderer   $assertionRenderer   The assertion renderer to use.
      * @param AssertionRecorder   $assertionRecorder   The assertion recorder to use.
      * @param Invoker             $invoker             The invoker to use.
@@ -43,6 +45,7 @@ class InstanceHandle extends AbstractHandle
         stdClass $state,
         StubFactory $stubFactory,
         StubVerifierFactory $stubVerifierFactory,
+        EmptyValueFactory $emptyValueFactory,
         AssertionRenderer $assertionRenderer,
         AssertionRecorder $assertionRecorder,
         Invoker $invoker
@@ -89,6 +92,7 @@ class InstanceHandle extends AbstractHandle
             $mock,
             $stubFactory,
             $stubVerifierFactory,
+            $emptyValueFactory,
             $assertionRenderer,
             $assertionRecorder,
             $invoker