+++ test/suite-generators/Stub/StubDataWithGeneratorsTest.php
@@ -33,7 +33,8 @@ class StubDataWithGeneratorsTest extends PHPUnit_Framework_TestCase
         $this->matcherVerifier = new MatcherVerifier();
         $this->invoker = new Invoker();
         $this->invocableInspector = new InvocableInspector();
-        $this->emptyValueFactory = new EmptyValueFactory();
+        $this->featureDetector = FeatureDetector::instance();
+        $this->emptyValueFactory = new EmptyValueFactory($this->featureDetector);
         $this->generatorAnswerBuilderFactory = GeneratorAnswerBuilderFactory::instance();
         $this->subject = new StubData(
             $this->callback,
@@ -138,8 +139,6 @@ class StubDataWithGeneratorsTest extends PHPUnit_Framework_TestCase
             $c = 'c';
             $d = 'd';
         };
-
-        $this->featureDetector = FeatureDetector::instance();
     }
 
     public function testGenerates()